refactor: implement domain-driven design architecture
