+++ src/application/mod.rs
@@ -1,0 +1,2 @@
+pub mod rss_summary;
+pub mod rss_summary_service;
