+++ src/mod.rs
@@ -1,3 +1,4 @@
+pub mod application;
+pub mod domain;
 pub mod infrastructure;
 pub mod model;
-pub mod rss_summary;
